Move existing tests to test/unit
